Namespacing the theme.
