+++ framework/classes/class-cartfront_homepage_control.php
@@ -5,100 +5,104 @@
  * @package cartfront
  */
 
-if ( ! defined( 'ABSPATH' ) ) {
-    exit;
-}
+namespace Niteo\WooCart\CartFront {
 
-if ( ! class_exists( 'Cartfront_Homepage_Control' ) ) :
-class Cartfront_Homepage_Control {
+    if ( ! defined( 'ABSPATH' ) ) {
+        exit;
+    }
 
-    /**
-     * The name of the hook for all the actions.
-     *
-     * @var     string
-     * @access  public
-     */
-    public $hook;
-
-    /**
-     * Constructor function.
-     *
-     * @access  public
-     * @since   1.0.0
-     */
-    public function __construct() {
-        $this->hook = (string) apply_filters( 'cf_hc_hook', 'homepage' );
+    if ( ! class_exists( 'Cartfront_Homepage_Control' ) ) :
+    class Cartfront_Homepage_Control {
 
         /**
-         * Re-structure components.
+         * The name of the hook for all the actions.
+         *
+         * @var     string
+         * @access  public
          */
-        if ( ! is_admin() ) {
-            add_action( 'get_header', array( &$this, 'restructure_components' ) );
+        public $hook;
+
+        /**
+         * Constructor function.
+         *
+         * @access  public
+         * @since   1.0.0
+         */
+        public function __construct() {
+            $this->hook = (string) apply_filters( 'cf_hc_hook', 'homepage' );
+
+            /**
+             * Re-structure components.
+             */
+            if ( ! is_admin() ) {
+                add_action( 'get_header', array( &$this, 'restructure_components' ) );
+            }
+
+            /**
+             * Initialize customizer.
+             */
+            $homepage_control_customizer = new Cartfront_Homepage_Control_Customizer();
         }
 
         /**
-         * Initialize customizer.
+         * Work through the stored data and display the components in the desired order, without the disabled components.
+         *
+         * @access  public
          */
-        $homepage_control_customizer = new Cartfront_Homepage_Control_Customizer();
-    }
+        public function restructure_components () {
+            $data       = get_theme_mod( 'cf_hc_data' );
+            $components = array();
 
-    /**
-     * Work through the stored data and display the components in the desired order, without the disabled components.
-     *
-     * @access  public
-     */
-    public function restructure_components () {
-        $data       = get_theme_mod( 'cf_hc_data' );
-        $components = array();
+            if ( isset( $data ) && '' != $data ) {
+                $components = explode( ',', $data );
 
-        if ( isset( $data ) && '' != $data ) {
-            $components = explode( ',', $data );
+                // Remove all existing actions on `homepage` hook.
+                remove_all_actions( $this->hook );
 
-            // Remove all existing actions on `homepage` hook.
-            remove_all_actions( $this->hook );
+                // Remove disabled components.
+                $components = $this->remove_disabled_items( $components );
 
-            // Remove disabled components.
-            $components = $this->remove_disabled_items( $components );
-
-            // Re-order the components.
-            if ( 0 < count( $components ) ) {
-                $count = 5;
-                foreach ( $components as $k => $v ) {
-                    if ( false !== strpos( $v, '@' ) ) {
-                        $obj_v = explode( '@' , $v );
-                        if ( class_exists( $obj_v[0] ) && method_exists( $obj_v[0], $obj_v[1] ) ) {
-                            add_action( $this->hook, array( $obj_v[0], $obj_v[1] ), $count );
+                // Re-order the components.
+                if ( 0 < count( $components ) ) {
+                    $count = 5;
+                    foreach ( $components as $k => $v ) {
+                        if ( false !== strpos( $v, '@' ) ) {
+                            $obj_v = explode( '@' , $v );
+                            if ( class_exists( $obj_v[0] ) && method_exists( $obj_v[0], $obj_v[1] ) ) {
+                                add_action( $this->hook, array( $obj_v[0], $obj_v[1] ), $count );
+                            }
+                        } else {
+                            if ( function_exists( $v ) ) {
+                                add_action( $this->hook, esc_attr( $v ), $count );
+                            }
                         }
-                    } else {
-                        if ( function_exists( $v ) ) {
-                            add_action( $this->hook, esc_attr( $v ), $count );
-                        }
+
+                        $count + 5;
                     }
-
-                    $count + 5;
-                }
-            }
-        }
-    }
-
-    /**
-     * Maybe remove disabled items from the main ordered array.
-     *
-     * @access  private
-     * @param   array $components   Array with components order.
-     * @return  array               Re-ordered components with disabled components removed.
-     */
-    private function remove_disabled_items( $components ) {
-        if ( 0 < count( $components ) ) {
-            foreach ( $components as $k => $v ) {
-                if ( false !== strpos( $v, '[disabled]' ) ) {
-                    unset( $components[ $k ] );
                 }
             }
         }
 
-        return $components;
+        /**
+         * Maybe remove disabled items from the main ordered array.
+         *
+         * @access  private
+         * @param   array $components   Array with components order.
+         * @return  array               Re-ordered components with disabled components removed.
+         */
+        private function remove_disabled_items( $components ) {
+            if ( 0 < count( $components ) ) {
+                foreach ( $components as $k => $v ) {
+                    if ( false !== strpos( $v, '[disabled]' ) ) {
+                        unset( $components[ $k ] );
+                    }
+                }
+            }
+
+            return $components;
+        }
+
     }
+    endif;
 
 }
-endif;
