+++ framework/include/sanitize.php
@@ -5,40 +5,48 @@
  * @package cartfront
  */
 
-/**
- * Sanitize input for the multi-select control.
- */
-if ( ! function_exists( 'cartfront_sanitize_multiselect' ) ) :
-function cartfront_sanitize_multiselect( $input ) {
-	if ( is_array( $input ) ) {
-		foreach ( $input as $key => $value ) {
-			$input[$key] = sanitize_text_field( $value );
-		}
-	} else {
-		$input = '';
+namespace Niteo\WooCart\CartFront {
+
+	if ( ! defined( 'ABSPATH' ) ) {
+	    exit;
 	}
 
-	return $input;
-}
-endif;
-
-/**
- * Sanitize input for the repeater control.
- */
-if ( ! function_exists( 'cartfront_sanitize_repeater' ) ) :
-function cartfront_sanitize_repeater( $input ) {
-	$input_decoded = json_decode( $input, true );
-
-	if ( ! empty( $input_decoded ) ) {
-		foreach ( $input_decoded as $boxk => $box ) {
-			foreach ( $box as $key => $value ) {
-				$input_decoded[$boxk][$key] = wp_kses_post( force_balance_tags( $value ) );
+	/**
+	 * Sanitize input for the multi-select control.
+	 */
+	if ( ! function_exists( 'sanitize_multiselect' ) ) :
+	function sanitize_multiselect( $input ) {
+		if ( is_array( $input ) ) {
+			foreach ( $input as $key => $value ) {
+				$input[$key] = sanitize_text_field( $value );
 			}
+		} else {
+			$input = '';
 		}
 
-		return json_encode( $input_decoded );
+		return $input;
 	}
+	endif;
 
-	return $input;
+	/**
+	 * Sanitize input for the repeater control.
+	 */
+	if ( ! function_exists( 'sanitize_repeater' ) ) :
+	function sanitize_repeater( $input ) {
+		$input_decoded = json_decode( $input, true );
+
+		if ( ! empty( $input_decoded ) ) {
+			foreach ( $input_decoded as $boxk => $box ) {
+				foreach ( $box as $key => $value ) {
+					$input_decoded[$boxk][$key] = wp_kses_post( force_balance_tags( $value ) );
+				}
+			}
+
+			return json_encode( $input_decoded );
+		}
+
+		return $input;
+	}
+	endif;
+
 }
-endif;
